pyproject.toml
[tool.flake8]
max-line-length = 140
ignore = ["E203", "W503"]
per-file-ignores = [
    '__init__.py:F401',
]

[tool.black]
line-length = 140

[tool.isort]
py_version = 39
line_length = 140

[tool.poetry]
name = "espeak-bot"
version = "0.1.0"
description = ""
authors = ["Andrii Yerko <aerko12@gmail.com>"]
readme = "README.md"
packages = [{include = "espeak_bot"}]

[tool.poetry.dependencies]
python = "^3.10"
python-telegram-bot = { version = "^20.0a4", allow-prereleases = true }
unidecode = "^1.3.6"


[tool.poetry.group.dev.dependencies]
pre-commit = "^2.20.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
